Add files via upload
The robot reads from an excel file the input date concerning the destination of multiple people: location, period, number of persons.  Using this data, it will search different available offers on the site https://www.booking.com/index.html and save in a new excel file the most expensive, respectively the cheapest 2 offers it found on the first two pages. The results will be sent on the email address which is also mentioned in the input excel file.